requirements.txt
argparse==1.2.1
certifi==2017.7.27.1
chardet==3.0.4
idna==2.6
nxt-python==2.2.2
requests==2.18.4
six==1.11.0
socketIO-client-2==0.7.5
urllib3==1.22
websocket-client==0.44.0
wsgiref==0.1.2
